+++ havister/templates/havister/index.html
@@ -12,7 +12,7 @@
 
 {% block nav-link %}
   {% if user.is_active %}
-    <a class="btn btn-success btn-xm" href="{% url 'user' user.username %}" role="button">{{ user.username|upper }}</a>
+    <a class="btn btn-success btn-xm" href="{% url 'users:index' user.username %}" role="button">{{ user.username|upper }}</a>
     {% else %}
     <a class="btn btn-success btn-xm" href="{% url 'login' %}" role="button">Login</a>
   {% endif %}
