--- players/templates/players/returns.html
@@ -1,62 +1,0 @@
-{% extends "base.html" %}
-
-{% block title %}
-<title>Players - Pointnut</title>
-{% endblock %}
-
-{% block nav-title %}
-<a class="mr-2" href="{% url 'players:index' %}"><i class="fas fa-chevron-left"></i></a>
-<span class="text-uppercase">{{ player.username }}</span>
-{% endblock %}
-
-{% block nav-link %}
-{% endblock %}
-
-{% block main %}
-<!-- Content Main -->
-<section class="box">
-  <div class="box-pad d-flex align-items-center">
-    <h4 class="my-1 text-darkgray mr-auto">{{ player.first_name }}</h4>
-    <p class="mb-0 text-capitalize">
-    {% for group in group_list %}
-      <span class="badge badge-green font-weight-normal">{{ group.name }}</span>
-    {% endfor %}
-    </p>
-  </div>
-  <!-- Section Body -->
-  {% if trade_list %}
-  <div class="box">
-    {% for trade in trade_list %}
-    <div class="box-pad d-flex align-items-center border-top">
-      <div class="d-none d-sm-block w-25 text-center pr-3">
-        <p class="h3 font-weight-normal text-green">{{ forloop.counter }}</p>
-      </div>
-      <div class="box">
-        <div class="d-flex align-items-center">
-          <h5 class="text-kr text-uppercase mr-auto">{{ trade }}</h5>
-          <p class="h5 text-{% if trade.change >= 0 %}orange{% else %}cyan{% endif %}">
-            <span>{{ trade.change|stringformat:"+0.2f" }}%</span>
-          </p>
-        </div>
-        <div class="d-flex">
-          <p class="mb-0 small text-muted mr-auto">
-            <span>{{ trade.date_opened|date:'Y.m.d' }}</span>
-            <span>~</span>
-            <span>{{ trade.date_closed|date:'Y.m.d' }}</span>
-          </p>
-          <p class="mb-0 small">
-            <span class="text-muted">{% if trade.position == 'L' %}LONG{% else %}SHORT{% endif %}</span>
-          </p>
-        </div>
-      </div>
-    </div>
-    {% endfor %}
-  </div>
-  {% else %}
-  <div class="box-pad">
-    <p>거래 기록이 없습니다.</p>
-  </div>
-  {% endif %}
-
-</section>
-{% endblock %}
