+++ templates/registration/password_change_done.html
@@ -6,7 +6,7 @@
 {% endblock %}
 
 {% block nav-title %}
-<a class="mr-2 green" href="{% url 'user' user.username %}"><i class="fas fa-chevron-left"></i></a>
+<a class="mr-2 green" href="{% url 'users:index' user.username %}"><i class="fas fa-chevron-left"></i></a>
 <span>Settings</span>
 {% endblock %}
 
@@ -27,7 +27,7 @@
 <section class="box-pad">
   <div class="my-2">
     <p>패스워드가 성공적으로 변경되었습니다.</p>
-    <a class="btn btn-success px-4" href="{% url 'user' user.username %}" role="button">확인</a>
+    <a class="btn btn-success px-4" href="{% url 'users:index' user.username %}" role="button">확인</a>
   </div>
 </section>
