+++ templates/registration/password_change_form.html
@@ -6,11 +6,26 @@
 {% endblock %}
 
 {% block nav-title %}
-<a class="mr-2 green" href="/"><i class="fas fa-chevron-left"></i></a>
+<a class="mr-2 green" href="{% url 'users:index' user.username %}"><i class="fas fa-chevron-left"></i></a>
 <span>Settings</span>
 {% endblock %}
 
 {% block nav-link %}
+<div class="dropdown">
+  <button class="btn btn-success btn-xm dropdown-toggle" type="button" id="dropdownMenuButton" data-toggle="dropdown" aria-haspopup="true" aria-expanded="false">
+    <i class="fas fa-user"></i>
+  </button>
+  <div class="dropdown-menu dropdown-menu-right" aria-labelledby="dropdownMenuButton">
+    <a class="dropdown-item disabled" href="#" tabindex="-1" aria-disabled="true">Profile</a>
+    <div class="dropdown-divider"></div>
+    <a class="dropdown-item disabled" href="#" tabindex="-1" aria-disabled="true">Havister</a>
+    <a class="dropdown-item disabled" href="#" tabindex="-1" aria-disabled="true">Trades</a>
+    <a class="dropdown-item" href="{% url 'users:returns' user.username %}">Returns</a>
+    <div class="dropdown-divider"></div>
+    <a class="dropdown-item active" href="#">Settings</a>
+    <a class="dropdown-item" href="{% url 'logout' %}">Logout</a>
+  </div>
+</div>
 {% endblock %}
 
 {% block header %}
